Vérification et suppression des fichiers inutiles
