--- animation-logo/styles.css
@@ -1,24 +1,0 @@
-.image-container {
-  overflow: hidden;
-  position: relative;
-  width: 100vw; /* Largeur de la fenêtre du navigateur */
-}
-
-footer{
-  position: fixed;
-  bottom: 0;
-}
-.img1 {
-  width: 5%;
-  height: auto;
-  transform: translateX(0); /* Point de départ */
-  transition: transform 2s linear; /* Durée et type de transition */
-}
-
-/* Lorsque vous déclencherez l'animation */
-.img2 {
-  width: 5%;
-  height: auto;
-  transform: translateX(100vw) scaleX(-1); /* Point de départ (à droite) et effet miroir horizontal */
-  transition: transform 2s linear; /* Durée et type de transition */
-}
