--- animation-logo/test.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Titre de la page</title>
-  <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-  <div class="image-container">
-    <img id="animation-button" class = "img1" src="logo.png" alt="Image à animer">
-  </div>
-  <footer>
-    <div class="image-container">
-        <img id="animation-button1" class = "img2" src="logo.png" alt="Image à animer">
-    </div>
-  </footer>
-  <script src="index.js"></script>
-</body>
-</html>
