+++ flask/tuto-flask/.gitignore
@@ -1,2 +1,3 @@
 /__pycache__
 tuto/__pycache__
+.gitignore
